+++ test_unstructured/file_utils/test_filetype.py
@@ -32,10 +32,13 @@ XLSX_MIME_TYPES = [
         ("example.jpg", FileType.JPG),
         ("fake-text.txt", FileType.TXT),
         ("fake-email.eml", FileType.EML),
-        ("unsupported/factbook.xml", FileType.XML),
+        ("factbook.xml", FileType.XML),
         ("example-10k.html", FileType.HTML),
         ("fake-html.html", FileType.HTML),
         ("stanley-cups.xlsx", FileType.XLSX),
+        # NOTE(robinson) - currently failing in the docker tests because the detected
+        # MIME type is text/csv
+        # ("stanley-cups.csv", FileType.CSV),
         ("fake-power-point.pptx", FileType.PPTX),
         ("winter-sports.epub", FileType.EPUB),
         ("spring-weather.html.json", FileType.JSON),
@@ -55,10 +58,11 @@ def test_detect_filetype_from_filename(file, expected):
         ("example.jpg", FileType.JPG),
         ("fake-text.txt", FileType.TXT),
         ("fake-email.eml", FileType.EML),
-        ("unsupported/factbook.xml", FileType.XML),
+        ("factbook.xml", FileType.XML),
         ("example-10k.html", FileType.HTML),
         ("fake-html.html", FileType.HTML),
         ("stanley-cups.xlsx", FileType.XLSX),
+        ("stanley-cups.csv", FileType.CSV),
         ("fake-power-point.pptx", FileType.PPTX),
         ("winter-sports.epub", FileType.EPUB),
         ("fake-doc.rtf", FileType.RTF),
@@ -88,12 +92,15 @@ def test_detect_filetype_from_filename_with_extension(monkeypatch, file, expecte
         ("example.jpg", FileType.JPG),
         ("fake-text.txt", FileType.TXT),
         ("fake-email.eml", FileType.EML),
-        ("unsupported/factbook.xml", FileType.XML),
+        ("factbook.xml", FileType.XML),
         # NOTE(robinson) - For the document, some operating systems return
         # */xml and some return */html. Either could be acceptable depending on the OS
         ("example-10k.html", [FileType.HTML, FileType.XML]),
         ("fake-html.html", FileType.HTML),
         ("stanley-cups.xlsx", FileType.XLSX),
+        # NOTE(robinson) - currently failing in the docker tests because the detected
+        # MIME type is text/csv
+        # ("stanley-cups.csv", FileType.CSV),
         ("fake-power-point.pptx", FileType.PPTX),
         ("winter-sports.epub", FileType.EPUB),
     ],
@@ -117,6 +124,12 @@ def test_detect_xml_application_xml(monkeypatch):
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake.xml")
     filetype = detect_filetype(filename=filename)
     assert filetype == FileType.XML
+
+
+def test_detect_text_csv(monkeypatch, filename="sample-docs/stanley-cup.csv"):
+    monkeypatch.setattr(magic, "from_file", lambda *args, **kwargs: "text/csv")
+    filetype = detect_filetype(filename=filename)
+    assert filetype == FileType.CSV
 
 
 def test_detect_xml_application_rtf(monkeypatch):