+++ unstructured/file_utils/filetype.py
@@ -67,6 +67,7 @@ class FileType(Enum):
     RTF = 41
     TXT = 42
     JSON = 43
+    CSV = 44
 
     # Markup Types
     HTML = 50
@@ -92,6 +93,12 @@ STR_TO_FILETYPE = {
     "image/jpeg": FileType.JPG,
     "image/png": FileType.PNG,
     "text/plain": FileType.TXT,
+    "text/x-csv": FileType.CSV,
+    "application/csv": FileType.CSV,
+    "application/x-csv": FileType.CSV,
+    "text/comma-separated-values": FileType.CSV,
+    "text/x-comma-separated-values": FileType.CSV,
+    "text/csv": FileType.CSV,
     "text/markdown": FileType.MD,
     "text/x-markdown": FileType.MD,
     "application/epub": FileType.EPUB,
@@ -111,9 +118,17 @@ STR_TO_FILETYPE = {
     "application/vnd.ms-outlook": FileType.MSG,
 }
 
-FILETYPE_TO_MIMETYPE = {
-    v: k for k, v in STR_TO_FILETYPE.items() if k not in ("text/x-markdown", "application/epub+zip")
-}
+MIMETYPES_TO_EXCLUDE = [
+    "text/x-markdown",
+    "application/epub+zip",
+    "text/x-csv",
+    "application/csv",
+    "application/x-csv",
+    "text/comma-separated-values",
+    "text/x-comma-separated-values",
+]
+
+FILETYPE_TO_MIMETYPE = {v: k for k, v in STR_TO_FILETYPE.items() if k not in MIMETYPES_TO_EXCLUDE}
 
 EXT_TO_FILETYPE = {
     ".pdf": FileType.PDF,
@@ -139,6 +154,7 @@ EXT_TO_FILETYPE = {
     ".epub": FileType.EPUB,
     ".msg": FileType.MSG,
     ".odt": FileType.ODT,
+    ".csv": FileType.CSV,
     None: FileType.UNK,
 }