+++ unstructured/partition/text.py
@@ -34,6 +34,8 @@ def partition_text(
     text: Optional[str] = None,
     encoding: Optional[str] = None,
     paragraph_grouper: Optional[Callable[[str], str]] = None,
+    metadata_filename: Optional[str] = None,
+    include_metadata: bool = True,
 ) -> List[Element]:
     """Partitions an .txt documents into its constituent elements.
     Parameters
@@ -49,6 +51,8 @@ def partition_text(
     paragrapher_grouper
         A str -> str function for fixing paragraphs that are interrupted by line breaks
         for formatting purposes.
+    include_metadata
+        Determines whether or not metadata is included in the output.
     """
     if text is not None and text.strip() == "" and not file and not filename:
         return []
@@ -72,22 +76,31 @@ def partition_text(
 
     file_content = split_by_paragraph(file_text)
 
+    metadata_filename = metadata_filename or filename
+
     elements: List[Element] = []
-    metadata = ElementMetadata(filename=filename)
+    metadata = (
+        ElementMetadata(filename=metadata_filename) if include_metadata else ElementMetadata()
+    )
     for ctext in file_content:
         ctext = ctext.strip()
 
-        if ctext == "":
-            continue
-        if is_bulleted_text(ctext):
-            elements.append(ListItem(text=clean_bullets(ctext), metadata=metadata))
-        elif is_us_city_state_zip(ctext):
-            elements.append(Address(text=ctext, metadata=metadata))
-        elif is_possible_narrative_text(ctext):
-            elements.append(NarrativeText(text=ctext, metadata=metadata))
-        elif is_possible_title(ctext):
-            elements.append(Title(text=ctext, metadata=metadata))
-        else:
-            elements.append(Text(text=ctext, metadata=metadata))
+        if ctext:
+            element = element_from_text(ctext)
+            element.metadata = metadata
+            elements.append(element)
 
     return elements
+
+
+def element_from_text(text: str) -> Element:
+    if is_bulleted_text(text):
+        return ListItem(text=clean_bullets(text))
+    elif is_us_city_state_zip(text):
+        return Address(text=text)
+    elif is_possible_narrative_text(text):
+        return NarrativeText(text=text)
+    elif is_possible_title(text):
+        return Title(text=text)
+    else:
+        return Text(text=text)